Added MIDI Controller Input Functionality
Now you are able to change the notes of points by pressing the corresponding key on your midi controller. The listener installed with this change only listens for NoteOn or NoteOff signals.
I also added a keybinding for the spacebar while focusing on the main canvas; this will start/stop the sequence.