General Feature Improvements and Code Cleaning
Changed the way points are modified in-context: encapsulated changes to the CC.nouspoints context rather than passing around variables.
Made sure that the canvas context is being returned to queue a draw instead of the points object
Changed the way points are drawn, to be more efficient and clear. Carets and paths are drawn within each point and not separately. This should avoid drawing problems and speed issues.
When points are deleted, the delete_path_to and delete_path_from functions are now automatically called.
When points are deleted, the path color now correctly reflects the path mode.
Download the update now by pulling changed files, or wait for 1.1.0, wherein these changes will be included.