New Feature: Muting
You can now mute points! Just select a point and push 'M' or select the Mute property and type true. This mostly avoids situations like portals where you may not want a note to play. The logic is unchanged, and new NoteSenders are even created, but if the mute property sent to the NoteSender is false, then no midi is sent. This allows information to continue to be sent across the sequence, in cases where you want a relative pitch to transfer, for example.